Add basic hgweb image with phabricator auth.
This change introduces a docker image which runs hgweb on uWSGI. This
hgweb instance support push over HTTP when authenticated using
phabricator credentials.

Phabricator credentials are expected to be a username and API key
provided through basic auth. These credentials are then checked in
a wsgi app wrapper before the request is passed to hgweb's wsgi
app.

Mozilla's version-control-tools repository is cloned as part of the
image to allow running extensions developed in that repository.